refactored code and added test cases

+++ com/freshworks/java/service/CRDService.java
@@ -22,6 +22,9 @@ public class CRDService {
         keyValueStorage.put(key, jsonData);
         String convertedData = gson.toJson(keyValueStorage);
         try {
+            File file = new File(filePath);
+            if (!file.isDirectory() || filePath.length() > 0)
+                filePath = "Data.txt";
             FileUtils.writeStringToFile(new File(filePath), convertedData);
             long timeStamp = System.currentTimeMillis() + TTL;
             writeCsv(key, filePath, timeStamp, true, masterCsvData);
@@ -87,19 +90,27 @@ public class CRDService {
     public static void deleteData(String key, Map<String, String> keyValueStorage, List<Master> masterCsvData) {
         keyValueStorage = CRDService.readCsv(key, masterCsvData);
         keyValueStorage.remove(key);
-        int temp = masterCsvData.size();
+        String convertedData = gson.toJson(keyValueStorage);
         List<Master> masterData = masterCsvData.stream().filter(master -> !master.getKey().equalsIgnoreCase(key)).collect(Collectors.toList());
         if (masterData.size() == 0) {
             new File("DataStore.csv").delete();
+            new File(masterCsvData.get(0).getFileLocation()).delete();
         } else
             masterData.forEach(master -> {
                 CRDService.writeCsv(master.getKey(), master.getFileLocation(), master.getTimestamp(), false, masterCsvData);
             });
         masterCsvData.stream().filter(master -> master.getKey().equalsIgnoreCase(key)).collect(Collectors.toList()).forEach(master -> {
-            new File(master.getFileLocation()).delete();
+            if(null != convertedData)
+                try {
+                    FileUtils.writeStringToFile(new File(master.getFileLocation()), convertedData);
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            else
+                new File(master.getFileLocation()).delete();
             masterCsvData.remove(master);
         });
-        if(temp == masterCsvData.size()) {
+        if(null != keyValueStorage.get(key)) {
             System.out.println("Deletion Unsuccessful!");
         } else
             System.out.println("Deleted Successfully");
